JobOutput: extract JobOutputPane
